Fixed sprite rendering (tex vertices are separate from screen vertices)
