+++ src/engine/Entity.java
@@ -1,4 +1,4 @@
-package game;
+package engine;
 
 import graphics.IRenderable;
