+++ src/engine/IUpdatable.java
@@ -1,4 +1,4 @@
-package game;
+package engine;
 
 public interface IUpdatable {
 	public void update(float deltaTime);
